refactor(e2e-suite): simplify Playwright detection in desktop
