+++ packages/suite-desktop-core/src/libs/process-switches.ts
@@ -18,7 +18,6 @@ export type SuiteSwitch =
     | 'log-no-print'
     | 'remove-user-data-on-start'
     | 'expose-connect-ws'
-    | 'expose-store'
     | 'state'; // very special handling, see `./app-utils.ts`
 
 /**
