+++ packages/suite-desktop-core/src/preload.ts
@@ -4,7 +4,6 @@ import { exposeIpcProxy } from '@trezor/ipc-proxy';
 import { getDesktopApi } from '@trezor/suite-desktop-api';
 
 import '@sentry/electron/preload'; // With this only IPCMode.Classic is ever taken into account
-import { hasSwitch } from './libs/process-switches';
 
 contextBridge.exposeInMainWorld(
     ...exposeIpcProxy(ipcRenderer, [
@@ -17,6 +16,3 @@ contextBridge.exposeInMainWorld(
 
 const desktopApi = getDesktopApi(ipcRenderer);
 contextBridge.exposeInMainWorld('desktopApi', desktopApi);
-contextBridge.exposeInMainWorld('desktopFlags', {
-    exposeStore: hasSwitch('expose-store'),
-});
