+++ packages/suite-desktop-ui/src/global.d.ts
@@ -2,5 +2,4 @@ type TrezorConnectIpcChannel = (method: string, ...params: any[]) => Promise<any
 
 interface Window {
     TrezorConnectIpcChannel?: TrezorConnectIpcChannel; // Electron API
-    desktopFlags?: { exposeStore?: boolean };
 }
