--- packages/suite-web/src/global.d.ts
@@ -1,6 +1,0 @@
-interface Window {
-    // Needed for Cypress and Playwright
-    Playwright?: any;
-    TrezorConnect?: any;
-    store?: any;
-}
