+++ packages/suite/global.d.ts
@@ -6,6 +6,12 @@ import { AnyAction } from '@suite-common/redux-utils';
 declare global {
     interface Window {
         __REDUX_DEVTOOLS_EXTENSION_COMPOSE__?: typeof compose;
+        // Flag that Suite is running in Playwright environment, but currently available only on Web
+        Playwright?: boolean;
+        // Internal global variable of Playwright, signifies reliably that Suite is running withing Playwright
+        playwright?: object;
+        // Needed for Playwright
+        store?: any;
     }
 }
