+++ packages/suite/src/hooks/suite/index.ts
@@ -21,6 +21,7 @@ export { useDebugLanguageShortcut } from './useDebugLanguageShortcut';
 export { useDisplayMode } from './useDisplayMode';
 export { useDefaultAccountLabel } from './useDefaultAccountLabel';
 export { useFirmwareInstallationProgressCheck } from './useFirmwareInstallationProgressCheck';
+export { usePlaywright, isRunningWithinPlaywright } from './usePlaywright';
 
 // replaced in suite-native
 export { useLocales } from 'src/hooks/suite/useLocales';
