+++ packages/suite/src/hooks/suite/usePlaywright.ts
@@ -1,24 +1,25 @@
 import { useEffect } from 'react';
 
-import TrezorConnect from '@trezor/connect';
-
 import { useStore } from 'src/hooks/suite/useStore';
+
+export const isRunningWithinPlaywright =
+    typeof window !== 'undefined' &&
+    // either of the properties can be present, depending on the Playwright environment
+    (window.Playwright || !!window.playwright);
 
 /**
  * Utility for running tests in Playwright.
- * Used to augment window object with redux store and TrezorConnect instance to make it accessible in tests
+ * Used to augment window object with redux store to make it accessible in tests.
  */
 export const usePlaywright = () => {
     const store = useStore();
 
     useEffect(() => {
-        if (typeof window !== 'undefined' && window.Playwright) {
+        if (isRunningWithinPlaywright) {
             window.store = store;
-            window.TrezorConnect = TrezorConnect;
 
             return () => {
                 delete window.store;
-                delete window.TrezorConnect;
             };
         }
     }, [store]);
